feat: Add detuning to tof experiment
